--- src/views/TmpEstablishment.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import './TmpEstablishment.css';
-import VideoChat from '../components/VideoChat';
-
-const TmpEstablishment = () => {
-  return (
-    <div className="app">
-      <header>
-        <h1>Video Chat with Hooks</h1>
-      </header>
-      <main>
-        <VideoChat />
-      </main>
-    </div>
-  );
-};
-
-export default TmpEstablishment;
